Added SplashActivity with Lottie animation (loading simulation) - added shared element transition between fragments - code cleaning - tmdb models parsing bugfix

+++ data/src/main/kotlin/io/mateam/playground2/data/dataSource/local/entity/MovieDbModel.kt
@@ -26,7 +26,7 @@ data class MovieDbModel(
     @TypeConverters(ProductionCompanyListConverter::class)
     val production_companies: List<DbProductionCompany>?,
     val release_date: String?,
-    val revenue: Int?,
+    val revenue: Long?,
     val runtime: Int?,
     @TypeConverters(SpokenLanguageListConverter::class)
     val spoken_languages: List<DbSpokenLanguage>?,
